v.surf.rst: point2d_parallel.c, function, test updates
