+++ vector/v.surf.rst/testsuite/test_vsurfrst.py
@@ -49,7 +49,9 @@ class TestVsurfrst(TestCase):
                 cls.tcurvature,
                 cls.mcurvature,
                 cls.deviations,
+                cls.deviations_threads,
                 cls.cvdev,
+                cls.cvdev_threads,
                 cls.treeseg,
                 cls.overwin,
             ],
